Feature: The SDL V1 screen has been introduced. And color is now uint32_t instead of int in this library.

+++ sc_point_iterator.h
@@ -20,8 +20,8 @@ struct line_point_iter {
 	struct point step;
 	struct point delta;
 	struct point acc;
-	int distance;
-	int count;
+	uint16_t distance;
+	uint16_t count;
 };
 
 struct rect_point_iter {
